+++ package.json
@@ -34,6 +34,7 @@
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
     "@types/react-router-dom": "^5.1.3",
+    "@types/react-select": "^3.0.10",
     "@typescript-eslint/eslint-plugin": "^2.17.0",
     "@typescript-eslint/parser": "^2.17.0",
     "autoprefixer": "^9.7.3",
@@ -92,6 +93,7 @@
     "react-dom": "^16.12.0",
     "react-grid-system": "^6.0.7",
     "react-player": "^1.15.2",
-    "react-router-dom": "^5.1.2"
+    "react-router-dom": "^5.1.2",
+    "react-select": "^3.0.8"
   }
 }
