+++ src/components/DiscoverySearch/DiscoverySearch.scss
@@ -1,3 +1,4 @@
 .discovery-search {
     background-color: theme-color("white");
+    padding: 40px 20px;
 }
