+++ src/components/DiscoverySearch/DiscoverySearch.tsx
@@ -1,0 +1,26 @@
+import React, { FunctionComponent } from 'react';
+
+import InputDropdown from '~/components/InputDropdown';
+import Input from '~/components/Input';
+import SimpleButton from '../SimpleButton';
+
+import './DiscoverySearch.scss';
+
+const DiscoverySearch: FunctionComponent = () => {
+    const className = 'discovery-search';
+    return (
+        <div className={className}>
+            <div className={`${className}__input-dropdown`}>
+                <InputDropdown />
+            </div>
+            <div className={`${className}__input`}>
+                <Input />
+            </div>
+            <div className={`${className}__button`}>
+                <SimpleButton title="SEARCH" />
+            </div>
+        </div>
+    );
+};
+
+export default DiscoverySearch;
