+++ src/components/InputDropdown/InputDropdown.scss
@@ -1,0 +1,12 @@
+.input-dropdown {
+    display: flex;
+    flex-direction: column;
+
+    &__label {
+        font-family: $font-family-primary;
+        font-size: font-size("paragraph");
+        font-weight: $font-weight-fat;
+        line-height: $line-height-medium;
+        margin-bottom: 10px;
+    }
+}
