--- src/sections/DiscoverySearch/DiscoverySearch.tsx
@@ -1,17 +1,0 @@
-import React, { FunctionComponent } from 'react';
-import { Container, Row, Col } from 'react-grid-system';
-
-const DiscoverySearch: FunctionComponent = () => {
-    const className = 'discovery-search';
-    return (
-        <div className={className}>
-            <Container>
-                <Row>
-                    <Col></Col>
-                </Row>
-            </Container>
-        </div>
-    );
-};
-
-export default DiscoverySearch;
