+++ src/views/Discovery/Discovery.tsx
@@ -2,7 +2,6 @@ import React, { FunctionComponent, Fragment } from 'react';
 
 import PageLayout from '~/layouts/PageLayout';
 import DiscoveryHero from '~/sections/DiscoveryHero';
-import DiscoverySearch from '~/sections/DiscoverySearch';
 
 const Discovery: FunctionComponent = () => {
     const className = 'discovery';
@@ -12,7 +11,6 @@ const Discovery: FunctionComponent = () => {
                 content={
                     <Fragment>
                         <DiscoveryHero />
-                        <DiscoverySearch />
                     </Fragment>
                 }
             />
